Remove Func definition from Go 1.11
