--- js/sync_test.go
@@ -1,11 +1,0 @@
-// Copyright 2019 The GopherWasm Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-// +build !wasm go1.12
-
-package js_test
-
-func isFuncSync() bool {
-	return true
-}
